% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitfile2synapse.R
\name{knitfile2synapse}
\alias{knitfile2synapse}
\title{Allow users to leverage knit3 when constructing Synapse Wiki content}
\usage{
knitfile2synapse(file, owner, parentWikiId = NULL, wikiName = NULL,
  overwrite = FALSE)
}
\arguments{
\item{owner}{A Synapse object which will own the resulting WikiPage (usually a Project, Folder, or File)}

\item{parentWikiId}{If the resulting WikiPage is to be a subpage of another WikiPage, this is the id for the parent WikiPage (NOTE: owner is still required)}

\item{wikiName}{A title for the resulting WikiPage - will default to the file name without the .Rmd extension}

\item{overwrite}{Only if owner specified and parentWikiId is NULL - flag for whether or not to overwrite the previous root WikiPage (if it exists)}

\item{File}{path to a local .Rmd file which to knit}
}
\value{
a WikiPage object as defined in the synapseClient
}
\description{
Allow users to leverage knit3 when constructing Synapse Wiki content
}

