% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitfile2synapse.R
\name{storeAndKnitToFileEntity}
\alias{storeAndKnitToFileEntity}
\title{storeAndKnitToFileEntity}
\usage{
storeAndKnitToFileEntity(file, parentId, fileName = NULL, owner = NULL,
  parentWikiId = NULL, wikiName = NULL, overwrite = FALSE,
  knitmd = TRUE, ...)
}
\arguments{
\item{file}{path to a local .Rmd file which to knit}

\item{parentId}{A synapseClient::Project or synapseClient::Folder entity (or Synapse ID of an entity) where the File will be created}

\item{fileName}{Name of the synapseClient::File to create}

\item{owner}{A Synapse entity (or Synapse ID of an entity) which will own the resulting WikiPage (usually a Project, Folder, or File)}

\item{parentWikiId}{If the resulting WikiPage is to be a subpage of another WikiPage, this is the ID for the parent WikiPage (NOTE: owner is still required)}

\item{wikiName}{A title for the resulting WikiPage - will default to the file name without the .Rmd extension}

\item{overwrite}{Only if owner specified and parentWikiId is NULL - flag for whether or not to overwrite the previous root WikiPage (if it exists). This will remove the history of changes for this Wiki page.}

\item{knitmd}{Flag for whether or not to knit; if FALSE and file already exists, don't knit it again}
}
\value{
a synapseClient::WikiPage object
}
\description{
Store a local RMarkdown file to Synapse and then knit it to that file's WikiPage.
}

